Compute vaccine coverage for a given value of alpha (percent of increase in final vaccine coverage)

Compute vaccine coverage for a given value of alpha (percent of increase in final vaccine coverage)

Compute events averted by increasing the coverage to a given value of alpha (percent of increase in final vaccine coverage)

Compute events averted by increasing the coverage to many values of alpha

Compute events averted by vaccination

Compute the number of individuals needed to vaccinate to prevent one event

Usage

compute_VC_alpha(diff_cumulative_coverage, target_coverage, alpha)

compute_events_avertable_by_increasing_coverage(
  events,
  diff_cumulative_coverage,
  target_coverage,
  alpha,
  vaccine_effectiveness
)

compute_events_avertable_by_increasing_coverage_many_alphas(
  events,
  diff_cumulative_coverage,
  target_coverage,
  alphas,
  vaccine_effectiveness
)

compute_events_averted_by_vaccination(
  events,
  cumulative_coverage,
  vaccine_effectiveness
)

compute_number_needed_to_vaccinate(
  events,
  events_averted,
  pop_at_risk,
  vaccine_effectiveness
)

Arguments

diff_cumulative_coverage

Vector of difference in cumulative coverage (cumulative_coverage_t - cumulative_coverage_t-1)

target_coverage

Target coverage (maximum coverage)

alpha

percent of increase in final vaccine coverage

events

Vector of number of events

vaccine_effectiveness

Vector of vaccine effectiveness

alphas

Vector of alpha values

cumulative_coverage

Vector of cumulative coverage

events_averted

Vector of number of events averted

pop_at_risk

Population at risk

Value

The vaccine coverage

The number of events averted

A data frame with the alpha values and the events averted

The number of events averted

The number of individuals needed to vaccinate to avert one event

Examples

diff_cumulative_coverage <- c(0, diff(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage))
target_coverage <- max(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage)
alpha <- 0.1
compute_VC_alpha(diff_cumulative_coverage, target_coverage, alpha)
#>  [1] 0.000000 0.000000 0.085500 0.123750 0.131625 0.108000 0.117000 0.070875
#>  [9] 0.073125 0.050625 0.039375 0.024750 0.021375 0.013500 0.012375 0.007875
#> [17] 0.006750 0.004500 0.003375 0.002250 0.001125 0.001125 0.001125 0.000000
#> [25] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [33] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [41] 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> [49] 0.000000 0.000000 0.000000 0.000000
events <- coverage_and_incidence_mock_data$incidence_data$events
diff_cumulative_coverage <- c(0, diff(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage))
target_coverage <- max(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage)
alpha <- 0.1
vaccine_effectiveness <- ve_mock_data$ve
compute_events_avertable_by_increasing_coverage(events, diff_cumulative_coverage,
target_coverage, alpha, vaccine_effectiveness)
#>  [1] 0.000000000 0.000000000 2.083148922 4.199830115 4.256792132 1.599197939
#>  [7] 0.965054815 1.103605549 0.512949020 0.373944572 0.231478586 0.130971005
#> [13] 0.026780188 0.028099463 0.039313806 0.019047362 0.003860255 0.002157079
#> [19] 0.004239641 0.002186660 0.000000000 0.000000000 0.000000000 0.000000000
#> [25] 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [31] 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [37] 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [43] 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> [49] 0.000000000 0.000000000 0.000000000 0.000000000
events <- coverage_and_incidence_mock_data$incidence_data$events
diff_cumulative_coverage <- c(0, diff(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage))
target_coverage <- max(coverage_and_incidence_mock_data$coverage_data$
cumulative_coverage)
alphas <- c(0.1, 0.2, 0.3)
vaccine_effectiveness <- ve_mock_data$ve
compute_events_avertable_by_increasing_coverage_many_alphas(events, diff_cumulative_coverage,
target_coverage, alphas, vaccine_effectiveness)
#>     alpha        nabe
#> 1     0.1 0.000000000
#> 2     0.1 0.000000000
#> 3     0.1 2.083148922
#> 4     0.1 4.199830115
#> 5     0.1 4.256792132
#> 6     0.1 1.599197939
#> 7     0.1 0.965054815
#> 8     0.1 1.103605549
#> 9     0.1 0.512949020
#> 10    0.1 0.373944572
#> 11    0.1 0.231478586
#> 12    0.1 0.130971005
#> 13    0.1 0.026780188
#> 14    0.1 0.028099463
#> 15    0.1 0.039313806
#> 16    0.1 0.019047362
#> 17    0.1 0.003860255
#> 18    0.1 0.002157079
#> 19    0.1 0.004239641
#> 20    0.1 0.002186660
#> 21    0.1 0.000000000
#> 22    0.1 0.000000000
#> 23    0.1 0.000000000
#> 24    0.1 0.000000000
#> 25    0.1 0.000000000
#> 26    0.1 0.000000000
#> 27    0.1 0.000000000
#> 28    0.1 0.000000000
#> 29    0.1 0.000000000
#> 30    0.1 0.000000000
#> 31    0.1 0.000000000
#> 32    0.1 0.000000000
#> 33    0.1 0.000000000
#> 34    0.1 0.000000000
#> 35    0.1 0.000000000
#> 36    0.1 0.000000000
#> 37    0.1 0.000000000
#> 38    0.1 0.000000000
#> 39    0.1 0.000000000
#> 40    0.1 0.000000000
#> 41    0.1 0.000000000
#> 42    0.1 0.000000000
#> 43    0.1 0.000000000
#> 44    0.1 0.000000000
#> 45    0.1 0.000000000
#> 46    0.1 0.000000000
#> 47    0.1 0.000000000
#> 48    0.1 0.000000000
#> 49    0.1 0.000000000
#> 50    0.1 0.000000000
#> 51    0.1 0.000000000
#> 52    0.1 0.000000000
#> 53    0.2 0.000000000
#> 54    0.2 0.000000000
#> 55    0.2 2.328794786
#> 56    0.2 4.713274483
#> 57    0.2 4.775868536
#> 58    0.2 1.791354938
#> 59    0.2 1.081951110
#> 60    0.2 1.233792948
#> 61    0.2 0.572911786
#> 62    0.2 0.417069331
#> 63    0.2 0.257935547
#> 64    0.2 0.145759021
#> 65    0.2 0.029800101
#> 66    0.2 0.031254153
#> 67    0.2 0.043720203
#> 68    0.2 0.021174939
#> 69    0.2 0.004291013
#> 70    0.2 0.002397329
#> 71    0.2 0.004711822
#> 72    0.2 0.002429917
#> 73    0.2 0.000000000
#> 74    0.2 0.000000000
#> 75    0.2 0.000000000
#> 76    0.2 0.000000000
#> 77    0.2 0.000000000
#> 78    0.2 0.000000000
#> 79    0.2 0.000000000
#> 80    0.2 0.000000000
#> 81    0.2 0.000000000
#> 82    0.2 0.000000000
#> 83    0.2 0.000000000
#> 84    0.2 0.000000000
#> 85    0.2 0.000000000
#> 86    0.2 0.000000000
#> 87    0.2 0.000000000
#> 88    0.2 0.000000000
#> 89    0.2 0.000000000
#> 90    0.2 0.000000000
#> 91    0.2 0.000000000
#> 92    0.2 0.000000000
#> 93    0.2 0.000000000
#> 94    0.2 0.000000000
#> 95    0.2 0.000000000
#> 96    0.2 0.000000000
#> 97    0.2 0.000000000
#> 98    0.2 0.000000000
#> 99    0.2 0.000000000
#> 100   0.2 0.000000000
#> 101   0.2 0.000000000
#> 102   0.2 0.000000000
#> 103   0.2 0.000000000
#> 104   0.2 0.000000000
#> 105   0.3 0.000000000
#> 106   0.3 0.000000000
#> 107   0.3 2.577470051
#> 108   0.3 5.237121057
#> 109   0.3 5.305163065
#> 110   0.3 1.986666912
#> 111   0.3 1.200974519
#> 112   0.3 1.365596560
#> 113   0.3 0.633502428
#> 114   0.3 0.460522352
#> 115   0.3 0.284544592
#> 116   0.3 0.160595013
#> 117   0.3 0.032829027
#> 118   0.3 0.034415422
#> 119   0.3 0.048134312
#> 120   0.3 0.023304770
#> 121   0.3 0.004722140
#> 122   0.3 0.002637695
#> 123   0.3 0.005184226
#> 124   0.3 0.002673234
#> 125   0.3 0.000000000
#> 126   0.3 0.000000000
#> 127   0.3 0.000000000
#> 128   0.3 0.000000000
#> 129   0.3 0.000000000
#> 130   0.3 0.000000000
#> 131   0.3 0.000000000
#> 132   0.3 0.000000000
#> 133   0.3 0.000000000
#> 134   0.3 0.000000000
#> 135   0.3 0.000000000
#> 136   0.3 0.000000000
#> 137   0.3 0.000000000
#> 138   0.3 0.000000000
#> 139   0.3 0.000000000
#> 140   0.3 0.000000000
#> 141   0.3 0.000000000
#> 142   0.3 0.000000000
#> 143   0.3 0.000000000
#> 144   0.3 0.000000000
#> 145   0.3 0.000000000
#> 146   0.3 0.000000000
#> 147   0.3 0.000000000
#> 148   0.3 0.000000000
#> 149   0.3 0.000000000
#> 150   0.3 0.000000000
#> 151   0.3 0.000000000
#> 152   0.3 0.000000000
#> 153   0.3 0.000000000
#> 154   0.3 0.000000000
#> 155   0.3 0.000000000
#> 156   0.3 0.000000000
compute_events_averted_by_vaccination(events = 100, cumulative_coverage = 0.5,
vaccine_effectiveness = 0.5)
#> [1] 33.33333
data("coverage_and_incidence_mock_data")
data("ve_mock_data")
events <- coverage_and_incidence_mock_data$incidence_data$events
cumulative_coverage <- coverage_and_incidence_mock_data$coverage_data$cumulative_coverage
vaccine_effectiveness <- ve_mock_data$ve
compute_events_averted_by_vaccination(events, cumulative_coverage, vaccine_effectiveness)
#>  [1]  0.0000000  0.0000000  1.8404774  6.6095710 11.0488282  7.3471758
#>  [7]  5.6472456 14.3417981  6.8704862  8.6083118  7.4468097  6.6467970
#> [13]  1.7373137  3.3178507  4.7765819  3.5559705  0.8146056  0.6098621
#> [19]  1.9862810  1.2618912  0.0000000  0.0000000  0.0000000  0.0000000
#> [25]  0.0000000  0.8982954  0.0000000  0.0000000  0.0000000  0.0000000
#> [31]  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> [37]  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> [43]  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000  0.0000000
#> [49]  0.0000000  0.0000000  0.0000000  0.0000000
events <- coverage_and_incidence_mock_data$incidence_data$events
cumulative_coverage <- coverage_and_incidence_mock_data$coverage_data$cumulative_coverage
events_averted <- compute_events_averted_by_vaccination(events,
cumulative_coverage, vaccine_effectiveness)
pop_at_risk <- 1000
vaccine_effectiveness <- ve_mock_data$ve
compute_number_needed_to_vaccinate(events, events_averted, pop_at_risk, vaccine_effectiveness)
#>  [1]   52.86934   47.56058   41.29363   28.14101   27.42372   54.30658
#>  [7]   89.06997   39.46507   91.84212   78.52875   95.47713  110.27868
#> [13]  432.85216  230.26955  162.24991  219.91183  967.33923 1298.65413
#> [19]  400.24548  631.59167         NA         NA         NA         NA
#> [25]         NA  890.57562         NA         NA         NA         NA
#> [31]         NA         NA         NA         NA         NA         NA
#> [37]         NA         NA         NA         NA         NA         NA
#> [43]         NA         NA         NA         NA         NA         NA
#> [49]         NA         NA         NA         NA